refactor(#143): разбить FeedScreen.kt (678 строк) на компоненты
FeedScreen.kt → 4 файла:
- FeedScreen.kt (180 строк) — корневой экран, координация состояний
- FeedHeader.kt (90 строк) — панель с городом, поиском, фильтром
- EventCard.kt (170 строк) — карточка события с избранным
- FeedContent.kt (241 строка) — DateStrip, ForYouSection, HorizontalEventRow
- EventImagePlaceholder.kt — вынесен в отдельный файл (используется из других пакетов)

Closes #143
