+++ composeApp/src/commonMain/kotlin/com/karrad/ticketsclient/ui/screen/feed/FeedContent.kt
@@ -1,0 +1,241 @@
+package com.karrad.ticketsclient.ui.screen.feed
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.rememberLazyListState
+import androidx.compose.foundation.pager.HorizontalPager
+import androidx.compose.foundation.pager.rememberPagerState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.derivedStateOf
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.karrad.ticketsclient.data.api.dto.DiscoveryFeedResponseDto
+import com.karrad.ticketsclient.data.api.dto.EventDto
+import kotlinx.datetime.Clock
+import kotlinx.datetime.DateTimeUnit
+import kotlinx.datetime.DayOfWeek
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.plus
+import kotlinx.datetime.toLocalDateTime
+
+@Composable
+internal fun FeedContent(
+    feed: DiscoveryFeedResponseDto,
+    selectedDay: Int,
+    onDaySelect: (Int) -> Unit,
+    onEventClick: (EventDto) -> Unit,
+    hasMore: Boolean = false,
+    onLoadMore: () -> Unit = {}
+) {
+    val listState = rememberLazyListState()
+    val reachedBottom by remember {
+        derivedStateOf {
+            val lastVisible = listState.layoutInfo.visibleItemsInfo.lastOrNull()
+            val total = listState.layoutInfo.totalItemsCount
+            lastVisible != null && total > 0 && lastVisible.index >= total - 2
+        }
+    }
+    LaunchedEffect(reachedBottom) {
+        if (reachedBottom && hasMore) onLoadMore()
+    }
+
+    LazyColumn(
+        state = listState,
+        modifier = Modifier.fillMaxSize().background(MaterialTheme.colorScheme.background),
+        contentPadding = PaddingValues(bottom = 96.dp)
+    ) {
+        stickyHeader {
+            DateStrip(selectedDay = selectedDay, onDaySelect = onDaySelect)
+        }
+
+        if (feed.forYou.isNotEmpty()) {
+            item {
+                SectionHeader("Для вас")
+                ForYouSection(events = feed.forYou, onEventClick = onEventClick)
+            }
+        }
+
+        feed.byCategory.forEach { entry ->
+            item {
+                SectionHeader(entry.category.label, hasMore = true)
+                HorizontalEventRow(events = entry.events, onEventClick = onEventClick)
+            }
+        }
+
+        if (hasMore) {
+            item {
+                Box(
+                    modifier = Modifier.fillMaxWidth().padding(16.dp),
+                    contentAlignment = Alignment.Center
+                ) {
+                    CircularProgressIndicator(
+                        modifier = Modifier.size(24.dp),
+                        strokeWidth = 2.dp,
+                        color = MaterialTheme.colorScheme.primary
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+internal fun FilteredResultsList(events: List<EventDto>, onEventClick: (EventDto) -> Unit) {
+    if (events.isEmpty()) {
+        Box(Modifier.fillMaxSize(), contentAlignment = Alignment.Center) {
+            Text(
+                "Ничего не найдено",
+                style = MaterialTheme.typography.bodyLarge,
+                color = MaterialTheme.colorScheme.onSurfaceVariant
+            )
+        }
+    } else {
+        LazyColumn(
+            modifier = Modifier.fillMaxSize().background(MaterialTheme.colorScheme.background),
+            contentPadding = PaddingValues(start = 16.dp, top = 8.dp, end = 16.dp, bottom = 96.dp),
+            verticalArrangement = Arrangement.spacedBy(12.dp)
+        ) {
+            items(events, key = { it.id }) { event ->
+                EventCard(event = event, cardWidth = null, onClick = { onEventClick(event) })
+            }
+        }
+    }
+}
+
+@Composable
+private fun DateStrip(selectedDay: Int, onDaySelect: (Int) -> Unit) {
+    val tz = TimeZone.currentSystemDefault()
+    val today = Clock.System.now().toLocalDateTime(tz).date
+    val listState = rememberLazyListState()
+
+    LazyRow(
+        state = listState,
+        modifier = Modifier
+            .fillMaxWidth()
+            .background(MaterialTheme.colorScheme.surface)
+            .padding(vertical = 10.dp),
+        contentPadding = PaddingValues(horizontal = 16.dp),
+        horizontalArrangement = Arrangement.spacedBy(6.dp)
+    ) {
+        items(14) { offset ->
+            val date = today.plus(offset, DateTimeUnit.DAY)
+            val isSelected = offset == selectedDay
+            Column(
+                modifier = Modifier
+                    .clip(RoundedCornerShape(12.dp))
+                    .clickable { onDaySelect(offset) }
+                    .background(if (isSelected) MaterialTheme.colorScheme.primary else Color.Transparent)
+                    .padding(horizontal = 10.dp, vertical = 6.dp),
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                Text(
+                    text = date.dayOfMonth.toString(),
+                    style = MaterialTheme.typography.titleSmall.copy(fontWeight = FontWeight.Bold),
+                    color = if (isSelected) Color.White else MaterialTheme.colorScheme.onBackground
+                )
+                Text(
+                    text = date.dayOfWeek.shortRu(),
+                    style = MaterialTheme.typography.labelSmall,
+                    color = if (isSelected) Color.White.copy(alpha = 0.8f) else MaterialTheme.colorScheme.onSurfaceVariant
+                )
+            }
+        }
+    }
+}
+
+private fun DayOfWeek.shortRu(): String = when (this) {
+    DayOfWeek.MONDAY    -> "ПН"
+    DayOfWeek.TUESDAY   -> "ВТ"
+    DayOfWeek.WEDNESDAY -> "СР"
+    DayOfWeek.THURSDAY  -> "ЧТ"
+    DayOfWeek.FRIDAY    -> "ПТ"
+    DayOfWeek.SATURDAY  -> "СБ"
+    DayOfWeek.SUNDAY    -> "ВС"
+}
+
+@Composable
+private fun SectionHeader(title: String, hasMore: Boolean = false) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 16.dp, vertical = 12.dp),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Text(title, style = MaterialTheme.typography.titleMedium.copy(fontWeight = FontWeight.Bold))
+        if (hasMore) {
+            Text(">", style = MaterialTheme.typography.bodyMedium, color = MaterialTheme.colorScheme.onSurfaceVariant)
+        }
+    }
+}
+
+@Composable
+private fun ForYouSection(events: List<EventDto>, onEventClick: (EventDto) -> Unit) {
+    val pagerState = rememberPagerState { events.size }
+    Column {
+        HorizontalPager(
+            state = pagerState,
+            contentPadding = PaddingValues(horizontal = 16.dp),
+            pageSpacing = 12.dp,
+            modifier = Modifier.fillMaxWidth()
+        ) { page ->
+            EventCard(event = events[page], cardWidth = null, imageHeight = 200.dp, onClick = { onEventClick(events[page]) })
+        }
+        Row(
+            modifier = Modifier.fillMaxWidth().padding(top = 10.dp),
+            horizontalArrangement = Arrangement.Center
+        ) {
+            repeat(events.size) { i ->
+                val active = pagerState.currentPage == i
+                Box(
+                    modifier = Modifier
+                        .padding(horizontal = 3.dp)
+                        .size(if (active) 8.dp else 6.dp)
+                        .clip(CircleShape)
+                        .background(
+                            if (active) MaterialTheme.colorScheme.primary
+                            else MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.3f)
+                        )
+                )
+            }
+        }
+        androidx.compose.foundation.layout.Spacer(Modifier.size(4.dp))
+    }
+}
+
+@Composable
+private fun HorizontalEventRow(events: List<EventDto>, onEventClick: (EventDto) -> Unit) {
+    LazyRow(
+        contentPadding = PaddingValues(horizontal = 16.dp),
+        horizontalArrangement = Arrangement.spacedBy(10.dp),
+        modifier = Modifier.padding(bottom = 4.dp)
+    ) {
+        items(events, key = { it.id }) { event ->
+            EventCard(event = event, cardWidth = 155.dp, imageHeight = 165.dp, onClick = { onEventClick(event) })
+        }
+    }
+}
