+++ composeApp/src/commonMain/kotlin/com/karrad/ticketsclient/ui/screen/feed/FeedHeader.kt
@@ -1,0 +1,90 @@
+package com.karrad.ticketsclient.ui.screen.feed
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Search
+import androidx.compose.material.icons.outlined.FilterList
+import androidx.compose.material.icons.outlined.LocationOn
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.karrad.ticketsclient.AppSession
+
+@Composable
+internal fun FeedHeader(
+    onSearchClick: () -> Unit = {},
+    onFilterClick: () -> Unit = {},
+    onCityClick: () -> Unit = {},
+    hasActiveFilter: Boolean = false
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .background(MaterialTheme.colorScheme.surface)
+            .padding(horizontal = 16.dp, vertical = 12.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Row(
+            modifier = Modifier
+                .clip(RoundedCornerShape(8.dp))
+                .clickable { onCityClick() }
+                .padding(horizontal = 4.dp, vertical = 2.dp),
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Icon(
+                Icons.Outlined.LocationOn,
+                contentDescription = "Сменить город",
+                tint = MaterialTheme.colorScheme.primary,
+                modifier = Modifier.size(18.dp)
+            )
+            Spacer(Modifier.width(4.dp))
+            Text(
+                text = AppSession.city,
+                style = MaterialTheme.typography.titleSmall.copy(fontWeight = FontWeight.SemiBold)
+            )
+        }
+        Spacer(Modifier.weight(1f))
+        IconButton(
+            onClick = onSearchClick,
+            modifier = Modifier
+                .size(36.dp)
+                .clip(CircleShape)
+                .background(MaterialTheme.colorScheme.surfaceVariant)
+        ) {
+            Icon(Icons.Default.Search, contentDescription = "Поиск",
+                tint = MaterialTheme.colorScheme.onSurface, modifier = Modifier.size(18.dp))
+        }
+        Spacer(Modifier.width(20.dp))
+        IconButton(
+            onClick = onFilterClick,
+            modifier = Modifier
+                .size(36.dp)
+                .clip(CircleShape)
+                .background(
+                    if (hasActiveFilter) MaterialTheme.colorScheme.primary
+                    else MaterialTheme.colorScheme.surfaceVariant
+                )
+        ) {
+            Icon(Icons.Outlined.FilterList, contentDescription = "Фильтры",
+                tint = if (hasActiveFilter) Color.White else MaterialTheme.colorScheme.onSurface,
+                modifier = Modifier.size(18.dp))
+        }
+    }
+}
